starting to clean up the new gym interface
