+++ grid2op/gym_compat/__init__.py
@@ -5,29 +5,19 @@
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
-__all__ = ["BaseGymAttrConverter", "GymEnv"]
+__all__ = ["BaseGymAttrConverter",
+           "GymEnv",
+           "GymObservationSpace",
+           "GymActionSpace",
+           "ScalerAttrConverter",
+           "MultiToTupleConverter",
+           "ContinuousToDiscreteConverter"]
 
 from grid2op.gym_compat.base_gym_attr_converter import BaseGymAttrConverter
 from grid2op.gym_compat.gymenv import GymEnv
 
-try:
-    from grid2op.gym_compat.gym_space_converter import GymObservationSpace, GymActionSpace
-    __all__.append("GymObservationSpace")
-    __all__.append("GymActionSpace")
-except ImportError:
-    # you must install open ai gym to benefit from this converter
-    pass
-
-try:
-    from grid2op.gym_compat.scaler_attr_converter import ScalerAttrConverter
-    __all__.append("ScalerAttrConverter")
-except ImportError:
-    # you must install open ai gym to benefit from this converter
-    pass
-
-try:
-    from grid2op.gym_compat.multi_to_tuple_converter import MultiToTupleConverter
-    __all__.append("MultiToTupleConverter")
-except ImportError:
-    # you must install open ai gym to benefit from this converter
-    pass
+from grid2op.gym_compat.gym_act_space import GymActionSpace
+from grid2op.gym_compat.gym_obs_space import GymObservationSpace
+from grid2op.gym_compat.scaler_attr_converter import ScalerAttrConverter
+from grid2op.gym_compat.multi_to_tuple_converter import MultiToTupleConverter
+from grid2op.gym_compat.continuous_to_discrete import ContinuousToDiscreteConverter
