+++ grid2op/gym_compat/gymenv.py
@@ -7,7 +7,8 @@
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
 import gym
-from grid2op.gym_compat.gym_space_converter import GymActionSpace, GymObservationSpace
+from grid2op.gym_compat.gym_obs_space import GymObservationSpace
+from grid2op.gym_compat.gym_act_space import GymActionSpace
 
 
 class GymEnv(gym.Env):
