+++ grid2op/tests/test_GymConverter.py
@@ -14,6 +14,8 @@ from grid2op.tests.helper_path_test import *
 from grid2op.MakeEnv import make
 from grid2op.Converter import IdToAct, ToVect
 from grid2op.gym_compat import GymActionSpace, GymObservationSpace
+from grid2op.gym_compat import GymEnv
+from grid2op.gym_compat import ContinuousToDiscreteConverter
 
 import pdb
 
@@ -60,7 +62,7 @@ class TestWithoutConverter(unittest.TestCase, BaseTestGymConverter):
             with make("l2rpn_wcci_2020", test=True) as env:
                 # test i can create
                 obs_space = GymObservationSpace(env)
-                act_space = GymActionSpace(env.action_space)
+                act_space = GymActionSpace(env)
 
     def test_json(self):
         with warnings.catch_warnings():
@@ -68,7 +70,7 @@ class TestWithoutConverter(unittest.TestCase, BaseTestGymConverter):
             with make("l2rpn_wcci_2020", test=True) as env:
                 # test i can create
                 obs_space = GymObservationSpace(env)
-                act_space = GymActionSpace(env.action_space)
+                act_space = GymActionSpace(env)
 
                 obs_space.seed(0)
                 act_space.seed(0)
@@ -101,7 +103,7 @@ class TestWithoutConverter(unittest.TestCase, BaseTestGymConverter):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
             with make("l2rpn_wcci_2020", test=True) as env:
-                act_space = GymActionSpace(env.action_space)
+                act_space = GymActionSpace(env)
 
                 act = env.action_space()
                 gym_act = act_space.to_gym(act)
@@ -130,7 +132,7 @@ class BaseTestConverter(BaseTestGymConverter):
             with make("l2rpn_wcci_2020", test=True) as env:
                 # test i can create
                 converter = self.init_converter(env)
-                act_space = GymActionSpace(converter)
+                act_space = GymActionSpace(env=env, converter=converter)
                 act_space.sample()
 
     def test_json(self):
@@ -139,7 +141,7 @@ class BaseTestConverter(BaseTestGymConverter):
             with make("l2rpn_wcci_2020", test=True) as env:
                 # test i can create
                 converter = self.init_converter(env)
-                act_space = GymActionSpace(converter)
+                act_space = GymActionSpace(env=env, converter=converter)
                 act_space.seed(0)
                 self._aux_test_json(act_space)
 
@@ -148,7 +150,7 @@ class BaseTestConverter(BaseTestGymConverter):
             warnings.filterwarnings("ignore")
             with make("l2rpn_wcci_2020", test=True) as env:
                 converter = self.init_converter(env)
-                act_space = GymActionSpace(converter)
+                act_space = GymActionSpace(env=env, converter=converter)
                 act_space.seed(0)
                 converter.seed(0)
 
@@ -187,3 +189,81 @@ class TestToVect(unittest.TestCase, BaseTestConverter):
 
     def setUp(self) -> None:
         BaseTestGymConverter.__init__(self)
+
+
+class TestDropAttr(unittest.TestCase):
+    """test the method to remove part of the attribute of the action / observation space"""
+    def test_keep_only_attr(self):
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            env = make("educ_case14_redisp", test=True)
+            gym_env = GymEnv(env)
+            attr_kept = sorted(("rho", "line_status", "actual_dispatch", "target_dispatch"))
+            ob_space = gym_env.observation_space
+            ob_space = ob_space.keep_only_attr(attr_kept)
+            assert np.all(sorted(ob_space.spaces.keys()) == attr_kept)
+
+    def test_ignore_attr(self):
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            env = make("educ_case14_redisp", test=True)
+            gym_env = GymEnv(env)
+            attr_deleted = sorted(("rho", "line_status", "actual_dispatch", "target_dispatch"))
+            ob_space = gym_env.observation_space
+            ob_space = ob_space.ignore_attr(attr_deleted)
+            for el in attr_deleted:
+                assert not el in ob_space.spaces
+
+
+class TestContinuousToDiscrete(unittest.TestCase):
+    """test the ContinuousToDiscreteConverter converter"""
+    def setUp(self) -> None:
+        self.tol = 1e-4
+
+    def test_split_in_3(self):
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            env = make("educ_case14_redisp", test=True)
+            gym_env = GymEnv(env)
+            act_space = gym_env.action_space
+            act_space = act_space.reencode_space("redispatch",
+                                                 ContinuousToDiscreteConverter(nb_bins=3,
+                                                                               init_space=act_space["redispatch"])
+                                                 )
+
+            # with 3 interval like [-10, 10] (the second generator)
+            # should be split => 0 -> [-10, -3.33), 1 => [-3.33, 3.33), [3.33, 10.]
+            # test the "all 0" action (all 0 => encoded to 1, because i have 3 bins)
+            g2op_object = np.array([0., 0., 0., 0., 0., 0.])
+            res = act_space._keys_encoding["_redispatch"].g2op_to_gym(g2op_object)
+            assert np.all(res == [1, 1, 0, 0, 0, 1])
+            res2 = act_space._keys_encoding["_redispatch"].gym_to_g2op(res)
+            assert np.all(res2 == 0.)
+
+            # test the all 0 action, but one is not 0 (negative)
+            g2op_object = np.array([0., -3.2, 0., 0., 0., 0.])
+            res = act_space._keys_encoding["_redispatch"].g2op_to_gym(g2op_object)
+            assert np.all(res == [1, 1, 0, 0, 0, 1])
+            res2 = act_space._keys_encoding["_redispatch"].gym_to_g2op(res)
+            assert np.all(res2 == 0.)
+
+            # test the all 0 action, but one is not 0 (positive)
+            g2op_object = np.array([0., 3.2, 0., 0., 0., 0.])
+            res = act_space._keys_encoding["_redispatch"].g2op_to_gym(g2op_object)
+            assert np.all(res == [1, 1, 0, 0, 0, 1])
+            res2 = act_space._keys_encoding["_redispatch"].gym_to_g2op(res)
+            assert np.all(res2 == 0.)
+
+            # test one is 2
+            g2op_object = np.array([0., 3.4, 0., 0., 0., 0.])
+            res = act_space._keys_encoding["_redispatch"].g2op_to_gym(g2op_object)
+            assert np.all(res == [1, 2, 0, 0, 0, 1])
+            res2 = act_space._keys_encoding["_redispatch"].gym_to_g2op(res)
+            # assert np.all(np.abs(res2 - [0., 6.66666, 0., 0., 0., 0.]) <= self.tol)
+
+            # test one is 0
+            g2op_object = np.array([0., -3.4, 0., 0., 0., 0.])
+            res = act_space._keys_encoding["_redispatch"].g2op_to_gym(g2op_object)
+            assert np.all(res == [1, 0, 0, 0, 0, 1])
+            res2 = act_space._keys_encoding["_redispatch"].gym_to_g2op(res)
+            # assert np.all(np.abs(res2 - [0., -6.66666, 0., 0., 0., 0.]) <= self.tol)
